Sprite flip enhanced - Started melee attack functionality
Sprite flip enhanced.

To correct some animation problems caused by irregularities in the size of the sprites. I add a transform property on each frame that needed to be corrected, but this transform need to be negative when the sprite is flipped, so I change the use of the flipX property in SpriteRender for use a negative scale and transform when the sprite need to be flipped.

This new functionality is more useful because let us add something in front of the character that will be flipped with him. For example a trigger to attack the enemies in front of the entity.